Remove some x64 warnings
git-svn-id: https://svn.code.sf.net/p/deinterlace/code/trunk/DScaler5@4670 6e12ab15-cf2f-4ffe-bf36-45892a6a5b89

+++ Filters/Decoders/MpegAudio/AudioDecoder.cpp
@@ -575,7 +575,7 @@ HRESULT CAudioDecoder::ProcessSample(IMediaSample* InSample, AM_SAMPLE2_PROPERTI
 
     m_Preroll = ((pSampleProperties->dwSampleFlags & AM_SAMPLE_PREROLL) == AM_SAMPLE_PREROLL);
 
-    int tmp = m_buff.size();
+    size_t tmp = m_buff.size();
     m_buff.resize(m_buff.size() + len);
     memcpy(&m_buff[0] + tmp, pDataIn, len);
     len += tmp;
